requirements.txt
pytest>=6.0.0
requests>=2.25.0
python-dotenv>=0.19.0